refactored to leverage common code between marathon and metronome around rpc calls
